Agentic_AI/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
SQLAlchemy==2.0.29
python-dotenv==1.0.1
langchain==0.2.3
langchain-community==0.2.3
langchain-huggingface==0.0.5
huggingface_hub==0.23.5
requests==2.31.0
rapidfuzz==3.8.1
pillow==10.3.0
python-multipart==0.0.9
streamlit==1.33.0
